Remove long double support
Long double isn't supported by valgrind (or CUDA for that matter), so
it's silly to have it anywhere in the codebase. This commit and the
changes to SoftEqual.test.cc  (which allow valgrind pass on emmet with
GCC 8.3 but fail on travis' VMs and GCC7) are being left in the history
for posterity.