Qualify Range types and functions for host and device
To help device/kernel function authors write cleaner loops, qualify
Range types and functions for host and device. The Range test is
extended to provide an optional CUDA component when this is available,
but is only stubbed in. It can be implemented once checks of PTX confirm
any cost to implementing loops this way on the device.