Disable but still compile HepMC3 and ROOT tests when unconfigured
This makes it clearer when running tests what functionality is disabled:
it shows "not run" rather than fewer tests. It also allows inclusion of
the header files even when the features are disabled.

The non-ROOT-specific IO routines have been moved into the main
celeritas library, and the library name changed to match.