Improve ROOT compatibility and ability to build without TPLs (#104)
* Add `CELER_NOT_CONFIGURED` macro
* Remove extra link requirement for event reader test
* Disable but still compile HepMC3 and ROOT tests when unconfigured
* Fix public/private and implementation include
* Add dockerfile script to launch local testing for an MR
* Add error checking code to ROOT calls
* Add necessary linker flags to force shared library link in Ubuntu

This makes it clearer when running tests what functionality is disabled:
it shows "not run" rather than fewer tests. It also allows inclusion of
the header files even when the features are disabled.

The non-ROOT-specific IO routines have been moved into the main
celeritas library, and the library name changed to match.