Add Pie helper classes (#131)
This new helper class is designed to simplify construction of device-compatible data and improve reusability across host and device.

The Pie class is templated on two new types: Ownership (which can be value, reference, or const reference) and Location (host or device). The generic Pie owns a Span under the hood, but the specialization for own/host has a std::vector and own/device has a DeviceVector. Templated assignment operators allow one set of operations to replace calls to copy_to_device, device_pointers, and host_pointers.