Use Collection for stack allocator backend (#169)
* Add helper algorithms and remove data accessor
* Refactor stack allocator to use Collection as backend
* Move thrust calls to separate cuda file for improved compile time

This eliminates the need for host allocator, storage class, etc. I
refactored the demo interactor detector stuff as part of it, as well as
some of the runner code that needed to be changed as a cosequence.

Additional changes were required in the livermore interactor since
the stack allocator now has a template parameter on memory space.
I refactored some parts of the livermore atomic relaxation to reduce
propagation of those template paramters and to improve program flow.