Set the lower applicability of models to zero
Cross section tables could extend below the lower applicable energy for a
model. This meant that a process could be sampled (with a nonzero cross
section) but not have a model that corresponds to that energy. Lowering the
model's limit to zero ensures that a model will always be selected for a
sampled process.