Add track initialization to demo loop (#251)
* Merge LDemoInterface with TrackInterface and update TrackInit test

* Fix initialization of track counters

* Initialize tracks from primaries and secondaries in demo loop

* Remove num_alive kernel and add max_steps to loop

The number of active tracks is already calculated in track initialization code,
so we didn't need to do it again in the demo-loop.

* Set correct interaction action for spawned, escaped, and boundary-crossing tracks

* Set the lower applicability of models to zero

Cross section tables could extend below the lower applicable energy for a
model. This meant that a process could be sampled (with a nonzero cross
section) but not have a model that corresponds to that energy. Lowering the
model's limit to zero ensures that a model will always be selected for a
sampled process.

* Move set_cuda_stack_size to Device.hh/cc and add to demo-loop

* Change brems interaction action from spawned to scattered

* Add bremsstrahlung to demo loop and a few fixes:

 - Fix references in kernel arguments
 - For the integral approach, allow the post-step cross section to be larger
   than the estimate of the maximum cross section over the step
 - Construct demo loop cutoffs with imported data

* Address reviewer feedback

* Update tests with some edge cases