Add Diagnostic interface with example (#262)
This commit adds the interface for collecting simulation and
stepping diagnostics. The interface allows one to obtain
details at the start and end of a simulation and its steps.

A simple example implementation of the `Diagnostic` interface
retrieves the number of per-thread particles, recorded at the end
of each step of `LDemoRun::run_gpu`.