Add algorithms for geometry tracking (#325)
* Rename device-friendly move/swap/forward to avoid conflict in algorithms

Cannot have `celeritas::swap` due to conflicts when `std` algorithms use unqualified `swap`

* Implement heap sort algorithm

* Add partition and generic lower_bound

* Combine stopwatch headers and add timing to simple unit tracker test

The change to double-precision for the stopwatch is because this is a CPU-only class (CPUs natively operate with double precision for the most part) and we may want the higher resolution (and there's no reason to switch this to single-precision with the rest of the physics code).

* Rename and refine move/forward/swap

Now move and forward are explicitly qualified where they're used. I've
also changed the sorting algorithm so that it uses `trivial_swap`
instead of `swap`: so it will now only apply to containers of trivial
types, but that's OK for now.