Add additional kernel occupancy stats to diagnostic
The CUDA `__launch_bounds__` takes `minBlocksPerMultiprocessor` (which I
think should be the same value as `OccupancyMaxActiveBlocksPerMultiprocessor`
returns if we always run the same number of threads per block?

HIP in contrast uses a `MIN_WARPS_PER_EU` argument to `__launch_bounds__`,
which can be calculated from the various occupancy statistics.