Add diagnostics and launch bounds (#380)
* Add additional device diagnostics

These include extra "useful to know" stats that aren't needed
as part of any potential runtime setup. It adds the "execution units
per multiprocessor" that HIP defines for AMD systems.

* Add additional kernel occupancy stats to diagnostic

The CUDA `__launch_bounds__` takes `minBlocksPerMultiprocessor` (which I
think should be the same value as `OccupancyMaxActiveBlocksPerMultiprocessor`
returns if we always run the same number of threads per block?

HIP in contrast uses a `MIN_WARPS_PER_EU` argument to `__launch_bounds__`,
which can be calculated from the various occupancy statistics.

* Make default block size configurable from environment

* Define CELER_EU_PER_MP and add more device props

* Unify celeritas-gen cmake scripts

* Add scripts for injecting launch bounds into kernel definitions

* Activate and adjust wildstyle launch bounds

* Inject manual V100-based launch bounds into initialize tracks

* Update MI250 generated launch bounds to match V100

* Change EU to MP since I think that's what HIP actually wants

Based on the feedback between the `__launch_bounds__` parameter and the
OccupancyMaxActiveBlocksPerMultiprocessor result, I think there's an
incorrect division by the number of EU per MP -- either that or the AMD
launch bounds are asking for the "CUDA EU" quantity rather than "native
EU" quantity.

* Disable launch bounds for vecgeom ndebug on CI

* Fix header and use absolute paths for generated files

* Auto-generate track initializer utils

* Revert "Change EU to MP since I think that's what HIP actually wants"

* Change MP to CU and add nomenclature description

* Change {block|grid|warp}_size to be foo_per_bar

This makes the translations and multiplications more intuitive.

* Reset launch bounds

* Fix EU calculation and typos, and add vendor output

* Update launch bounds with "default" values and update lb keys

* Update occupancies resulting from reduced thread count

* Revert launch bounds to compiler-calculated values

Disable CELERITAS_LAUNCH_BOUNDS as part of CI. We can return to this
later to try to tweak performance, but for now it doesn't seem to be
helping.