Reject discrete interaction when cross sections drop below validity
Currently the model applicability limits have to be uniform across materials,
which is inconsistent with cross section data. So we make sure the range extends
across all possible energies enclosing the cross sections, and rely on the
start-of-step XS calculation to reject the processes, and a new end-of-step XS
calculation to reject those that have gone out of range.