Refactor along-step for extensibility
The goal is to allow different along-step kernels based on the physics needs. There is now a generic "along-step" algorithm (function templated on other function arguments) that takes a combination of:
- multiple scattering (so we can create a specialized kernel without it, increasing occupancy)
- propagator (to allow different magnetic field types with arbitrary parameter data)
- energy loss applier (to allow for no-energy-loss neutral particles or faster simulation with eloss disabled by the user)

The along-step kernel has moved out of the physics class and is now simply an action. Further actions can be added. There are two example kernels in place:
- neutral (for photons: no magnetic field, no energy loss)
- along-step general (the current method basically: optional eloss, optional urban msc)

We can/should add more for "no fluctuations" and "uniform magnetic field" (for testing) but I wanted to wait for feedback on the basic interfaces.

Current deficiencies:
- The along-step classes have no easy way of being auto generated because they rely on different data and different on-device operators that use the data.
- I could move the UrbanMsc data out of the physics like I did with fluctuation, but it would require changing the importdata and such so I'm going to postpone that