rdc cmake: Workaround for detecting if RDC is needed for a library.
We currently do not properly detect if a library needs RDC or not.  Temporarily use
'vecgeom is being used' as a proxy for this.

In the long term we would need to improve mechanism (and factor code out) to delegate
the setting (and related creation) of STATIC/MIDDLE/FINAL until we discover a library
need to be be split including the cases:
        * contains cuda code and CUDA_SEPARABLE_COMPILATION set explicitly
        * contains cuda and dependent on RDC libraries
        * dependent of 2 independent RDC libraries
      all those so far can be detected by either add_library or target_link_library
      but the following is harder to intercept:
        * CUDA_SEPARABLE_COMPILATION set directly after the add_library