Initial test of "Celerino" pseudoparticle
Geant4 defines a "Geantino" pseudoparticle to assist in debugging
transportation. Both uncharged and charged versions are present,
the latter for interaction with fields. All are massless, spin-0 and
stable. Whilst Celeritas uses an in-depth unit test suite to debug
similar topics, discussions suggested that a similar "Celerino"
particle could be useful for work on transportation and fields.

Use demo-loop app as an initial testing ground for declaring and
running Celerinos:

- Add Celerino to import particle list after others have been created
  from Geant4 or ROOT.
- Create Celerino with same physics properties as Geantino.
- Declare Celerino's PDG code as 81, following PDG recommendation for
  internal MC pseudoparticles.
- Add tests of demo-loop using Celerino primaries

Original tests of demo-loop should still pass, whilst those using
Celerino primaries will fail as implemented in this first pass.