Support linking to Geant4 static libraries
If Geant4 is installed with only static libraries, the FindGeant4
wrapper will fail as it directly sets properties on the unavailable
Geant4::G4global shared target.

Check that the shared Geant4::G4global target exists before setting
its properties. Fallback to the static target if the shared does not
exist. This is safe against dual installs, as Geant4 will prefer
shared targets in this case.