Support linking to Geant4 static libraries (#623)
If Geant4 is installed with only static libraries, the FindGeant4
wrapper will fail as it directly sets properties on the unavailable
Geant4::G4global shared target.

Check that the shared Geant4::G4global target exists before setting
its properties. Fallback to the static target if the shared does not
exist. This is safe against dual installs, as Geant4 will prefer
shared targets in this case.

An install of Geant4 can contain both static and shared targets. Whilst
Celeritas only links to one set, the required compile feature should
be applied to each existing variant of Geant4::G4global. This permits
Celeritas to cleanly link to one or other variant if desired in future.