Add failing along-step test
The positron pre-step energy is near the discontinuity in the MSC positron
cross sections. In the true path --> geom path conversion in the MSC step limit
calculation, the final transport mfp is larger than the initial transport mfp,
and alpha is negative. In the MSC scatter geom path --> true path conversion
the incorrect transformation is used since alpha < 0. The calculated true path
length is too large and is clamped to the interaction length even though the
track has moved to a boundary. This triggers an assertion in along-step when
the mfp goes to zero.