Fix incorrect Urban MSC path length correction due to discontinuity in positron cross section (#642)
* Add failing along-step test

The positron pre-step energy is near the discontinuity in the MSC positron
cross sections. In the true path --> geom path conversion in the MSC step limit
calculation, the final transport mfp is larger than the initial transport mfp,
and alpha is negative. In the MSC scatter geom path --> true path conversion
the incorrect transformation is used since alpha < 0. The calculated true path
length is too large and is clamped to the interaction length even though the
track has moved to a boundary. This triggers an assertion in along-step when
the mfp goes to zero.

* Fix incorrect path length correction in Urban MSC

- Use exact comparison instead of checking for negative alpha in
  UrbanMscScatter::calc_true_path so that steps with negative alpha values
  due to the positron xs discontinuity will use the correct geom path --> true
  path transformation
- Update along-step test

* Move MSC alpha sentinel value to MscStep