Remove field driver convergence assertions
After discussion with Soon we think that these field driver failures are
likely due to low energy particles in unusual places, so possibly it's a
problem stemming from another part of the code (like failure to apply
cuts).

Calculating the step lengths with reduced accuracy is probably
acceptable in those cases, and certainly better than stopping the
simulation.