Use separate tables for LaTeX and HTML output
LaTeX does not render the restructured text table of physics very
well. Use directives to only use the restructured text table for
HTML, with a separate raw LaTeX table for that output.

Add multirow and threeparttable packages to LaTeX preamble to
assist in formatting the table like the HTML version.