Document available physics processes and models (#653)
* Document available physics processes and models

Provide basic tabulation of EM processes and models in Celeritas.
Take initial layout from that used in the G4HepEM documentation.
Applicability ranges taken from model's applicability() functions,
though this does not account for thresholds for secondary production.

Import Doxygen/Breathe for listed Celeritas classes to have a cross
reference in the API docs.

* Enable documentation for EM classes and fix UrbanMsc link

* Use implementation classes and fix some documentation

* Use separate tables for LaTeX and HTML output

LaTeX does not render the restructured text table of physics very
well. Use directives to only use the restructured text table for
HTML, with a separate raw LaTeX table for that output.

Add multirow and threeparttable packages to LaTeX preamble to
assist in formatting the table like the HTML version.

* Drop Geant4 equivalents in physics table

The API documentation for the interactors references the corresponding
Geant4 classes/documents describing the model/implementation. This
should be sufficient for cross-referencing and so we can drop the
Geant4 equivalent column in the overview physics table.

* Fix latex build

* Make sure breathe understands file extensions

* Use en dashes for numeric range

* Additional en dashing and name/layout corrections

* Link to Interactors rather than Models

The Interactor classes implement the actual physics/sampling and
are thus more relevant to users interested in this.

---------

Co-authored-by: Seth R Johnson <johnsonsr@ornl.gov>