Alternative implementation for filling with std::iota / thrust::sequence
The collection of ThreadId to TrackSlotId maps ThreadId to TrackSlotId::size_type instead.
When returning a track slot id, the CoreTrackView will wrap the index into a TrackSlotId.

Advantage of this variant is that we no longer need operator+(OpaqueId, OpaqueId) and operator*(OpaqueId, int)
which don't make sense for the pointer semantics of OpaqueId