Fix step limiter when losing all energy over a step (#780)
This assigns the "range" limiter when a particle loses all energy over a step (even when it's not marked as range-limited) and has no at-rest interaction.

The "discrete" process will be incorrectly selected otherwise. This isn't a problem currently because we also filter on track.status(), but we want the action ID to be consistent.