Only use ROOT in celer-g4 when Celeritas uses Geant4 (#798)
* Only use ROOT in celer-g4 when Celeritas uses Geant4

In the situation CELERITAS_USE_Geant4 is false and CELERITAS_USE_ROOT
is true, dictionary generation will be attempted. This relies on
use of Geant4 includes and thus compilation will fail.

Only generate a dictionary/use ROOT in celer-g4 when both ROOT and
Geant4 are used.

* Keep ROOT conditional separate from Geant4 to reduce indent

Retain behaviour of only generating dictionaries when both ROOT
and Geant4 are in use.