Refactor theta/phi calculation for G4Para/G4Trap into free function
Use of Geant4 prior to 11.0 requires the theta/phi angles for G4Para
and G4Trap solids to be calculated from the components of the symmetry
axis vector. The encoding of the components is identical for both and
can thus be refactored into a common function.

Refactor calculation of theta/phi angles into a free function in
the anonymous namespace. Take symmetry vector/axis as argument,
returning theta, phi as a pair.