Fix build errors from Geant4@10 (#830)
* Workaround no G4UIcommand::ConvertToLongInt in Geant4 < 10.7

Conversion from string to long int is only present in G4UIcommand from
Geant4 10.7, so to support 10.6 (e.g. for ATLAS) a workaround is
required.

Only use G4UIcommand::ConvertToLongInt when compiling against Geant 10.7
or newer. For older versions, use direct conversion via istringstream
as is done in the Geant4 implementation.

* Workaround G4Para/Trap accessors not available in Geant4 < 11

Direct accessors for theta/phi/alpha parameters on G4Para and G4Trap
were only added in Geant4 11.0, but we want to support 10.6 and newer
for ATLAS.

When compiling against Geant4 < 11, calculate theta, phi and alphas
using available symmetry axis and tan(alpha) accessors. Use direct
accessors otherwise.

* Put expected/normal code first in if preprocessor

Addresses review comments to have the "else" clause as the exceptional
case (here, for backward compatibility).

* Refactor theta/phi calculation for G4Para/G4Trap into free function

Use of Geant4 prior to 11.0 requires the theta/phi angles for G4Para
and G4Trap solids to be calculated from the components of the symmetry
axis vector. The encoding of the components is identical for both and
can thus be refactored into a common function.

Refactor calculation of theta/phi angles into a free function in
the anonymous namespace. Take symmetry vector/axis as argument,
returning theta, phi as a pair.