Document how to link loadable shared libraries with Celeritas
Integration of Celeritas in ATLAS's FullSimLight required it to
be used in and linked from a shared library loaded by the FullSimLight
application at runtime (a "plugin"). `celeritas_target_link_libraries`
on the plugin library did not result in the needed "_final" device code
library being linked to it, leading to undefined symbols at load time.
Traced to `celeritas_target_link_libraries` not considering a shared
library as an endpoint required full device link resolution.

Briefly describe this use case in documentation and provide example
of how to link a "plugin" shared library to Celeritas targets.