Expose SDSetupOptions "enabled" flag in SetupOptionsMessenger
In testing ATLAS integration it was found useful to switch SD callbacks
off/on whilst other parts of the integration were worked on and to test
functionality. Though a relatively simple switch in the wrapper around
SimpleOffload/SetupOptions, it is an obvious parameter for exposure as
a runtime option through SetupOptionsMessenger to help integration and
possibly later performance evaluation.

Add a new "enabled" command in options messenger coupled to the
SDSetupOptions enabled flag. Place command in "detector" subdirectory as
a pure naming marker of its purpose. Document it in a new table entry in
header file for Doxygen export.