Clarify use cases for linking to SHARED plugin library
Celeritas recommends use of celeritas_add_library to support the full
RDC chain but it may not be possible to use this in projects that
provide their own wrappers over the raw CMake commands.

Add clarification on the use case here, retaining example use of raw
CMake commands for simplest possible use case.