Rewrite soft equality for simplicity (#932)
* Rewrite soft equality as symmetric with fewer instructions but different inf behavior

* Add EqualOr functor and update unit test logic for infinities

* Update SoftEq to have a constexpr constructor and accessors

* Add justification for soft unit vector equality

* Fix narrowing casts in test

* fixup! Fix narrowing casts in test

* Fix tests for change in soft comparison against zero

I think the new behavior is more sensible. Now, the absolute tolerance
is compared when one of the values is zero, because we don't know a
priori whether zero is the correct value to have.

* Address momentum comparison failures

There are small differences in the momentum change that I can't explain,
but the energy balance is correct to machine precision so I think it's ok.

* Add reference unit for derivative constant test

* Use EXPECT_SOFT_EQ since the new version is much nicer