Refactor test name generation to be based on path
Previously the test subdirectory would get repeated by the
CELERITASTEST_PREFIX. Now that prefix is no longer necessary, since the
relative path becomes part of the filename and part of the test name.
The prefix is no longer repeated because the subdirectory is only added
to the "target name" and the "test name", not the "output name".