Give priority to existing environment variables for G4 data location (the app version)
This is necessary for proper operation when building and testing Celeritas
in a ATLAS build environment.  In the ATLAS build environment the Geant4
cmake fragment contains incorrect information about the location of the
data file (it points to the long gone build directory) so we need
to trust the environment variables instead.