Fix VecGeom middle library name.
`libvecgeomcuda.[a|so]` is the final library and it includes the result
of nvlink.  It is not appropriate to use as the intermediary link library.

VecGeom also never produce a shared 'middle' library, so we just always
need to use `libvecgeomcuda_static.a`

For the static build this is perfect.
For the shared build (especially in conjunction with user's own
package shared build), this is not working.  So we just keep
using libvecgeomcuda.so as the middle library hoping that the
nvlink result is not pickup at the end.