Fix linking Athena with VecGeom+CUDA (#1489)
* Fix VecGeom middle library name.

`libvecgeomcuda.[a|so]` is the final library and it includes the result
of nvlink.  It is not appropriate to use as the intermediary link library.

VecGeom also never produce a shared 'middle' library, so we just always
need to use `libvecgeomcuda_static.a`

For the static build this is perfect.
For the shared build (especially in conjunction with user's own
package shared build), this is not working.  So we just keep
using libvecgeomcuda.so as the middle library hoping that the
nvlink result is not pickup at the end.

* Use new VecGeom 'middle' library if available.

This is available in VecGeom 1.2.10 and later and is noticeable also by the setting
of the CUDA_RDC property

* cuda_rdc_target_link_libraries: fix handling of static library dependency.

The previous code was inaccurately looking at the target rather than its dependency when making the choice on how
to place the library

* CudaRdc: add/fix support for private dependency.

They are record as '$<LINK_ONLY:(.*)>' which was not unbox to replace the cuda libraries
with their 'middle' counterpart.