Improve CUDA_RUNTIME_LIBRARY determination (for export libs)
In case a library does not have a CUDA_RUNTIME_LIBRARY
(i.e. case of exported library with CudaRdcUtils v4 or older)
we now 'guess' the type by checking whether the middle and
object library are the same, if they are this is a static library
and it "probably" require the static CUDA runtime.