Always explictly add cuda runtime to executable.
If the executable source uses the CUDA API only in C++ files,
cmake can not detect the need to add the CUDA runtime library.
However since the type of the runtime library (static vs shared)
depends on the time of the depend library, it is not easy for the
user to known which one to add.

For now, just always add the proper runtime library to the
executable runtime library.